Fixed bugs in Pong.
First, I moved from a simple format of the loss function to the
log-loss. I don't fully understand why log-loss is so cool, but I'm
duplicating Andrey Karpathy's code for now, so I went with what he had.

What really was the issue was that "A * B", when A is a vertical vector of
length N, and B is a horizontal one of length N, silently results in the outer
product - a matrix of size N x N. Which is very much not what we wanted; in
our case we then summed the entries in A * B to obtain the loss - and
unfortunately B was zero-mean, which led to A * B being zero mean.
Flattening both solved the issue.

After that, the gradients exploded. I presume Karpathy didn't have this
issue, because he didn't really instantiate the loss, he went directly
to the policy gradient. I'm wondering whether I could do the same; but
for now I'm doing a naive scenario where I have a loss term I
differentiate, so I used the standard trick of separating it from zero
(by adding a constant and scaling down). Also, I added an l2-loss term,
although it doesn't seem to actually do much.

At this point the network ran happily for 1.4K episodes so far; and
while the weights are still expanding, they're not doing it
exponentially, but rather in a somewhat sub-linear fashion, so I hope
it's going to be fine.